feat(ingredient): 재료 섭취 로직 분리
